Add placeholder node circle textures while images are loading.
Sidenote: site is now interactive and visually usable MUCH sooner in the load process than it used to be. (due to the changes made in this commit and the previous commit's loading changes).

Also finally removed all references to circleSprite. Fully transferred over to imgSprite. (the temp circle texture simply gets set as the imgSprite texture at 1st).